+++ src/mono/mono/mini/CMakeLists.txt
@@ -307,6 +307,7 @@ set(interp_sources
     interp/mintops.h
     interp/mintops.c
     interp/transform.c
+    interp/transform-opt.c
     interp/tiering.h
     interp/tiering.c
     interp/jiterpreter.c
